refactor(core): move model json utilities into shared
