+++ packages/core/src/ai-model/models/ui-tars/adapter.ts
@@ -1,7 +1,7 @@
 import { type TModelFamily, UITarsModelVersion } from '@midscene/shared/env';
 import { assert } from '@midscene/shared/utils';
 import type { ModelAdapterDefinition } from '../../model-adapter/types';
-import { parseModelResponseJson } from '../../service-caller/json';
+import { parseModelResponseJson } from '../../shared/json';
 import {
   type LocateResultValue,
   createLocateResultValue,
