+++ packages/core/src/ai-model/prompt/extraction.ts
@@ -1,6 +1,6 @@
 import type { AIDataExtractionResponse, ServiceExtractParam } from '@/types';
 import { getPreferredLanguage } from '@midscene/shared/env';
-import { parseModelResponseJson } from '../service-caller/json';
+import { parseModelResponseJson } from '../shared/json';
 import { extractXMLTag } from '../shared/xml';
 
 export function buildTypeQueryDemandValue(
